// src/api/models/userDetailItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { UserCarItem } from './userCarItem';

export interface UserDetailItem {
  /** ID */
  id: string;
  /** 이름 */
  name: string;
  /**
   * 이메일
   * @nullable
   */
  email?: string | null;
  /** 전화번호 */
  phoneNumber: string;
  /**
   * 주소
   * @nullable
   */
  address?: string | null;
  /**
   * 탈퇴 이유
   * @nullable
   */
  deleteReason?: string | null;
  /** 추천인 코드 */
  referralCode: string;
  /**
   * 회원이 등록한 추천인 코드
   * @nullable
   */
  referrerCode?: string | null;
  /** 생성일 */
  createdAt: string;
  /**
   * 삭제일
   * @nullable
   */
  deletedAt?: string | null;
  /**
   * 정지일
   * @nullable
   */
  bannedAt?: string | null;
  /** 삭제 여부 */
  isDeleted: boolean;
  /** 정지 여부 */
  isBanned: boolean;
  /** 마케팅 수신 여부 */
  isMarketing: boolean;
  /** 등록차량 목록 */
  cars: UserCarItem[];
}